backend/requirements.txt
fastapi
fastapi-mcp
uvicorn
python-dotenv
pinecone
nemoguardrails
langchain
langchain-groq
groq
pydantic
httpx
pandas
numpy
tiktoken
python-multipart
tavily-python
asyncio
datasets